backend/requirements.txt
# ── Core framework ────────────────────────────────────────────────
annotated-types==0.7.0
anyio==4.11.0
fastapi==0.110.1
h11==0.16.0
starlette==0.37.2
uvicorn==0.25.0
slowapi==0.1.9
limits==3.6.0

# ── Database ──────────────────────────────────────────────────────
dnspython==2.8.0
motor==3.3.1
pymongo==4.5.0

# ── Data validation ───────────────────────────────────────────────
pydantic==2.12.4
pydantic_core==2.41.5
typing_extensions==4.15.0
typing-inspection==0.4.2
email-validator==2.3.0

# ── Auth & security ───────────────────────────────────────────────
bcrypt==4.1.3
cffi==2.0.0
cryptography==46.0.3
ecdsa==0.19.1
oauthlib==3.3.1
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
PyJWT==2.10.1
python-jose==3.5.0
rsa==4.9.1

# ── AWS / storage ─────────────────────────────────────────────────
boto3==1.40.76
botocore==1.40.76
jmespath==1.0.1
s3transfer==0.14.0

# ── Email ─────────────────────────────────────────────────────────
resend==2.19.0

# ── HTTP & networking ─────────────────────────────────────────────
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
requests==2.32.5
requests-oauthlib==2.0.0
urllib3==2.5.0

# ── Utilities ─────────────────────────────────────────────────────
click==8.3.1
python-dotenv==1.2.1
python-multipart==0.0.20
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
packaging==25.0
six==1.17.0
sniffio==1.3.1

# ── CLI / display ─────────────────────────────────────────────────
rich==14.2.0
Pygments==2.19.2
markdown-it-py==4.0.0
mdurl==0.1.2
shellingham==1.5.4
typer==0.20.0
pytokens==0.3.0
watchfiles==1.1.1
